Squashed commit, updates to v1.4:
- add TPM2 support
- add package whitelisting ASPs and demo for userspace APB
- resolves potential memory leaks found through static analysis
- updated README.md
- removed ROADMAP.md
- add Layered Attestaion demo
- fix potential null dereferences found by static analysis
- added ASAN to CI; fixed all ASAN issues
- resolve some uncharacterized issues found by static analysis
- fixes basic selinux tests
- improvements/cleanup of RPM build and test in CI
- removes uneeded appraiser permissions
- documentation changes
- adds test for contract creation and verification ASPs
- changed decrypt API in order to remove unneeded argument
- added test for encrypt and decrypt ASP
- removed dead code
- decrypt ASP changes to remove uneeded calls, labels, and fixed up dlog call
- astyle formatting of changed files
- fixes errors in SELinux policy, refactores code slightly to improve security posture of APB
- corrects type in selinux policy
- changes required to get layered attester and appraiser to work
- adds support for dynamic measurement request address space variables
- fix ups to compress and encrypt ASPs
- added ASPs for receiving measurement contracts for the layered attestation APB
- adds measurement spec for layered attestation APB to makefile
- corrects typos in the measurement spec
- fixes serialization of the dynamic_measurement_request_address_space
- updates selinux policy for several APBs
- adds Docker demos
- minor changes regarding the place registration capability
- update version to v1.4

+++ src/types/address_space/kernel_as.c
@@ -191,7 +191,7 @@ static gboolean kernel_address_equal(const address *a, const address *b)
 static guint kernel_address_hash(const address *a)
 {
     struct kernel_address *ka = (struct kernel_address *)a;
-    return ka->kaddr;
+    return (guint) ka->kaddr;
 }
 
 static void *kernel_read_bytes(address *a, size_t size)
